dév. shomgt4 sprint1 en cours

+++ bo/viewtiff.php
@@ -1,77 +1,146 @@
 <?php
 /* bo/viewtiff.php - Visualisation et validation d'une livraison utilisant georaster-layer-for-leaflet
- * Benoit DAVID - 11-19/7/2023
+ * Benoit DAVID - 11-22/7/2023
+ * paramètres GET
+ *  - path - chemin du répertoire contenant des 7z de cartes
+ *  - map  - nom de base du fichier 7z d'une carte (sans l'extension .7z)
+ *  - tif  - entrée du .tif dans l'archive 7z
+ *
+ * Faire des tests de viewtiff avec:
+ *  - 2 cartes normales standard sans cartouches
+ *  - 2 cartes normales standard avec cartouches
+ *  - toutes les cartes spéciales anciennes et nouvelles
+ *  - la carte normale mal géoréférencée
+ *  - toutes les cartes normales à cheval sur l'antiméridien
 */
+require_once __DIR__.'/login.inc.php';
 require_once __DIR__.'/mapmetadata.inc.php';
-require_once __DIR__.'/conform.php';
+require_once __DIR__.'/maparchive.php';
 
-// var url_to_geotiff_file = "http://localhost/shomgeotiff/incoming/20230710/6735/6735_pal300.tif";
 define ('HTML_HEAD', "<!DOCTYPE html>\n<html><head><title>viewtiff</title></head><body>\n");
 define ('JSON_OPTIONS', JSON_PRETTY_PRINT|JSON_UNESCAPED_SLASHES|JSON_UNESCAPED_UNICODE|JSON_THROW_ON_ERROR);
+define ('TEST_MAPS', [
+    "2 cartes normales sans cartouche" => [
+      'path=/incoming/20230710&map=6735' =>
+        "6735 - Pas de Calais - De Boulogne-sur-Mer à Zeebrugge, estuaire de la Tamise (Thames)",
+      'path=/archives/7441&map=7441-2009' => "7441-2009 - Abords et Ports de Monaco",
+    ],
+    "2 cartes normales avec partie principale et cartouches"=> [
+      'path=/archives/7594&map=7594-1903' => "7594-1903 - De la Pointe Ebba au Cap de la Découverte",
+      'path=/incoming/20230710&map=7090' => "7090 - De la Pointe de Barfleur à Saint-Vaast-la-Hougue",
+    ],
+    "2 cartes normales avec cartouches mais sans partie principale" => [
+      'path=/archives/7207&map=7207-2303'=> "7207-2303 - Ports de Fécamp et du Tréport",
+      'path=/archives/7427&map=7427-1724'=>
+        "7427-1724 - La Gironde- De Mortagne-sur-Gironde au Bec d'Ambès"
+        ." - La Garonne et La Dordogne jusqu'à Bordeaux et Libourne",
+    ],
+    "Carte 7620 mal géoréférencée" => [
+      'path=/archives/7620&map=7620-1903'=> "7620-1903 - Approches d'Anguilla",
+      'path=/archives/7620&map=7620-2242'=> "7620-2242 - Approches d'Anguilla",
+    ],
+    "Les anciennes cartes spéciales" => [
+      'path=/archives/7330&map=7330-1726'=> 
+        "7330-1726 - De Cherbourg à Hendaye - Action de l'Etat en Mer en Zone Maritime Atlantique",
+      'path=/archives/7344&map=7344-1726'=>
+        "7344-1726 - De Brest à la frontière belge - Action de l'Etat en Mer - Zone Manche et Mer du Nord",
+      'path=/archives/7360&map=7360-1726'=>
+        "7360-1726 - De Cerbère à Menton - Action de l'Etat en Mer - Zone Méditerranée",
+      'path=/archives/8101&map=8101-1726'=> "8101-1726 - MANCHEGRID - Carte générale",
+      'path=/archives/8502&map=8502-1726'=> "8502-1726 - Action de l'Etat en Mer en ZMSOI",
+      'path=/archives/8509&map=8509-1944'=> 
+        "8509-1944 - Action de l''Etat en Mer - Nouvelle-Calédonie - Wallis et Futun",
+      'path=/archives/8510&map=8510-1944'=> "8510-1944 - Délimitations des zones maritimes",
+      'path=/archives/8517&map=8517-1944'=>
+        "8517-1944 - Carte simplifiée de l''action de l''Etat en Mer des ZEE Polynésie Française et Clipperton",
+      'path=/archives/8523&map=8523-2224'=>
+        "8523-2224 - Carte d''Action de l'État en Mer - Océan Atlantique Nord - Zone maritime Antilles-Guyane",
+    ],
+    "Les nouvelles cartes spéciales" => [
+      'path=/incoming/20230628aem&map=7330'=> 
+        "7330 - De Cherbourg à Hendaye - Action de l'Etat en Mer en Zone Maritime Atlantique",
+      'path=/doublons/20230626&map=7344'=>
+        "7344 - Carte d’Action de l’État en Mer Zone Manche et Mer du Nord - \"De Brest à la frontière Belge\"",
+      'path=/doublons/20230626&map=7360'=>
+        "7360 - Carte d’Action de l’État en Mer Zone Méditerranée - \"De Cerbère à Menton\"",
+      'path=/attente/20230628aem&map=8502'=>
+        "8502 - Carte d’Action de l'État en Mer en Zone Maritime Sud de l'Océan Indien ZMSOI",
+      'path=/attente/20230628aem&map=8509'=>
+        "8509 - Carte d’Action de l’État en Mer - Nouvelle-Calédonie - Wallis et Futuna",
+      'path=/attente/20230628aem&map=8510'=> "8510 - Délimitation des zones maritimes",
+      'path=/attente/20230628aem&map=8517'=> 
+        "8517 - Carte simplifiée d’Action de l’État en Mer des ZEE Polynésie française et Clipperton",
+      'path=/attente/20230628aem&map=8523'=>
+        "8523 - Carte d’Action de l’État en Mer - Océan Atlantique Nord - Zone maritime Antilles-Guyane",
+    ],
+    "Cartes à cheval sur l'antiméridien" => [
+      'path=/archives/6835&map=6835-2311'=> "6835-2311 - Océan Pacifique Nord - Partie Est",
+      'path=/archives/6977&map=6977-2304'=> "6977-2304 - Océan Pacifique Nord - Partie Nord-Ouest",
+      /*
+      7021
+      7271
+      7166
+      6671
+      6670
+      6817
+      7283
+      */
+    ],
+    "Tests d'erreurs"=> [
+      'path=/attente/20230628aem&map=xx'=> "Le fichier n'existe pas",
+    ],
+  ]
+); // cartes de tests 
+
+
+if (!($login = Login::login())) {
+  die("Accès non autorisé\n");
+}
+
+if (!($PF_PATH = getenv('SHOMGT3_PORTFOLIO_PATH')))
+  throw new Exception("Variables d'env. SHOMGT3_PORTFOLIO_PATH non définie");
 
 if (!isset($_GET['path'])) { // affichage de la liste des livraisons 
   echo HTML_HEAD;
   $groups = [
-    'incoming'=> "Liste des livraisons",
-    'attente'=> "Liste des livraisons en attente",
-    'archives'=> "Liste des archives",
+    '/incoming'=> "Livraisons",
+    '/attente'=> "Livraisons en attente",
+    '/archives'=> "Archives",
   ];
-  if (($_GET['action'] ?? null) == 'clean') {
-    echo "<h2>Nettoyage</h2>\n";
-    foreach ($groups as $gname => $title) {
-      echo "$title:<ul>\n";
-      foreach (new DirectoryIterator(SHOMGEOTIFF.$gname) as $incoming) {
-        if (in_array($incoming, ['.','..','.DS_Store'])) continue;
-        foreach (new DirectoryIterator(SHOMGEOTIFF."$gname/$incoming") as $dirname) {
-          if (in_array($dirname, ['.','..','.DS_Store'])) continue;
-          if (is_dir(SHOMGEOTIFF."$gname/$incoming/$dirname")) {
-            echo "<li>Effacement du répertoire $gname/$incoming/$dirname<ul>\n";
-            foreach (new DirectoryIterator(SHOMGEOTIFF."$gname/$incoming/$dirname") as $filename) {
-              if (in_array($filename, ['.','..'])) continue;
-              if (substr($filename, -3) == '.7z') { // je n'efface pas les archives 7z 
-                echo "<li><b>Erreur sur $filename</b></li>\n";
-              }
-              else {
-                echo "<li>$filename</li>\n";
-                if (isset($_GET['confirm']))
-                  unlink(SHOMGEOTIFF."$gname/$incoming/$dirname/$filename");
-              }
-            }
-            if (isset($_GET['confirm']))
-              rmdir(SHOMGEOTIFF."$gname/$incoming/$dirname");
-            echo "</ul></li>\n";
-          }
-        }
-      }
-      echo "</ul>\n";
-    }
-    if (!isset($_GET['confirm']))
-      echo "<a href='?action=clean&confirm=true'>confirmer ?</a><br>\n";
-    die();
-  }
   foreach ($groups as $gname => $title) {
-    echo "$title:<ul>\n";
-    foreach (new DirectoryIterator(SHOMGEOTIFF.$gname) as $incoming) {
+    echo "<h3>$title</h3><ul>\n";
+    foreach (new DirectoryIterator($PF_PATH.$gname) as $incoming) {
       if (in_array($incoming, ['.','..','.DS_Store'])) continue;
       echo "<li><a href='?path=$gname/$incoming'>$incoming</a></li>\n";
     }
     echo "</ul>\n";
   }
-  echo "<a href='?action=clean'>Nettoyage</a><br>\n";
+  echo "<a href='?path=tests&action=tests'>Cartes de tests</a>\n";
   die();
 }
 
-if (!isset($_GET['map'])) { // affichage du contenu de la livraison 
+if (($_GET['action'] ?? null) == 'tests') {
+  echo HTML_HEAD,"<h3>Cartes de test</h3>\n";
+  foreach (TEST_MAPS as $gtitle => $group) {
+    echo "<h4>$gtitle</h4><ul>\n";
+    foreach ($group as $path => $title)
+      echo "<li><a href='?$path'>$title</a></li>\n";
+    echo "</ul>\n";
+  }
+  die();
+}
+
+if (!isset($_GET['map'])) { // affichage du contenu de la livraison ou du répertoire d'archives 
   echo HTML_HEAD;
   echo "<h2>Répertoire $_GET[path]</h2>\n";
-  if (substr($_GET['path'], 0, 8) <> 'archives')
+  if (substr($_GET['path'], 0, 9) <> '/archives')
     echo "<ul>\n";
   $first = true;
-  foreach (new DirectoryIterator(SHOMGEOTIFF.$_GET['path']) as $map) {
+  foreach (new DirectoryIterator($PF_PATH.$_GET['path']) as $map) {
     if (substr($map, -3) <> '.7z') continue;
     //echo "<li>"; print_r($md); echo "</li>";
-    if (substr($_GET['path'], 0, 8) == 'archives') { // cas d'une archive de carte
-      $md = MapMetadata::getFrom7z(SHOMGEOTIFF."$_GET[path]/$map");
+    if (substr($_GET['path'], 0, 9) == '/archives') { // cas d'une archive de carte
+      $md = MapMetadata::getFrom7z("$PF_PATH$_GET[path]/$map");
       if ($first) {
         echo $md['title'] ?? $map,"<ul>\n";
         $first = false;
@@ -80,7 +149,7 @@ if (!isset($_GET['map'])) { // affichage du contenu de la livraison
       echo "<li><a href='?path=$_GET[path]&map=$mapid'>",$md['edition'] ?? $mapid,"</li>\n";
     }
     else { // cas d'une livraison
-      $md = MapMetadata::getFrom7z(SHOMGEOTIFF."$_GET[path]/$map");
+      $md = MapMetadata::getFrom7z("$PF_PATH$_GET[path]/$map");
       $mapnum = substr($map, 0, -3);
       echo "<li><a href='?path=$_GET[path]&map=$mapnum'>",$md['title'] ?? $mapnum,"</a></li>\n";
     }
@@ -89,32 +158,45 @@ if (!isset($_GET['map'])) { // affichage du contenu de la livraison
   die();
 }
 
+if (!is_file("$PF_PATH$_GET[path]/$_GET[map].7z"))
+  die("Erreur le fichier $PF_PATH$_GET[path]/$_GET[map].7z n'existe pas\n");
+
 switch ($_GET['action'] ?? null) {
   case null: { // affichage des caractéristiques de la carte
     echo HTML_HEAD;
     MapCat::init();
-    $map = new Map(SHOMGEOTIFF.$_GET['path'], $_GET['map']);
-    $map->showAsHtml(SHOMGEOTIFF.$_GET['path'], $_GET['map'], MapCat::get(substr($_GET['map'], 0, 4)));
+    $mapNum = substr($_GET['map'], 0, 4);
+    $map = new MapArchive("$PF_PATH$_GET[path]/$_GET[map].7z", $mapNum, MapCat::get($mapNum));
+    $map->showAsHtml(MapCat::get(substr($_GET['map'], 0, 4)));
     die();
   }
   case 'gdalinfo': { // affichage du gdalinfo correspondant à un tif
-    $gdalinfo = new GdalInfo(SHOMGEOTIFF."$_GET[path]/$_GET[map]/$_GET[tif]");
+    $archive = new SevenZipArchive("$PF_PATH$_GET[path]/$_GET[map].7z");
+    $archive->extractTo(__DIR__.'/temp', $_GET['tif']);
+    $gdalinfo = new GdalInfo(__DIR__.'/temp/'.$_GET['tif']);
     header('Content-type: application/json; charset="utf-8"');
     echo json_encode($gdalinfo->asArray(), JSON_OPTIONS);
+    unlink(__DIR__.'/temp/'.$_GET['tif']);
+    rmdir(__DIR__.'/temp/'.dirname($_GET['tif']));
     die();
   }
   case 'viewtiff': { // affichage des tiff de la carte dans Leaflet
+    MapCat::init();
     $tifs = [];
-    $map = new Map(SHOMGEOTIFF.$_GET['path'], $_GET['map']);
+    $mapNum = substr($_GET['map'], 0, 4);
+    $map = new MapArchive("$PF_PATH$_GET[path]/$_GET[map].7z", $mapNum, MapCat::get($mapNum));
     foreach ($map->gtiffs() as $fileName) {
       echo "$fileName<br>\n";
-      $tifs[substr($fileName, 5, -4)] = "http://localhost/shomgeotiff/$_GET[path]/$_GET[map]/$fileName";
+      $shomgeotiffUrl = "$_SERVER[REQUEST_SCHEME]://$_SERVER[SERVER_NAME]".dirname($_SERVER['PHP_SELF'])."/shomgeotiff.php";
+      $tifs[substr($fileName, 5, -4)] = "$shomgeotiffUrl$_GET[path]/$_GET[map].7z/$fileName";
     }
     echo "<pre>tifs = "; print_r($tifs); echo "</pre>\n";
     $bounds = ($gbox = $map->gbox()) ? $gbox->latLngBounds() : [];
     echo "<pre>bounds = "; print_r($bounds); echo "</pre>\n";
-    if (!$tifs || !$gbox)
-      die("Affichage impossible\n");
+    if (!$tifs)
+      die("Affichage impossible car aucun GéoTiff à afficher\n");
+    if (!$bounds)
+      die("Affichage impossible car impossible de déterminer l'extension à afficher\n");
     //die("Ok");
     break;
   }
@@ -165,11 +247,7 @@ switch ($_GET['action'] ?? null) {
           parseGeoraster(arrayBuffer).then(georaster => {
             console.log("georaster:", georaster);
             overlays[<?php echo "'$name'"; ?>] = new GeoRasterLayer({georaster: georaster, opacity: 1.0, resolution: 256});
-<?php if ($name == 'pal300') { ?>
-            map.fitBounds(overlays['pal300'].getBounds());
-<?php } ?>
             map.addLayer(overlays[<?php echo "'$name'"; ?>]);
-            
 <?php } ?>
                   map.addLayer(baseLayers['OSM']);
                   L.control.layers(baseLayers, overlays).addTo(map);
